Extract KAPI VPA reconciliation into dedicated file
