+++ pkg/operation/botanist/component/kubeapiserver/kube_apiserver.go
@@ -24,7 +24,6 @@ import (
 
 	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
-	autoscalingv1 "k8s.io/api/autoscaling/v1"
 	autoscalingv2beta1 "k8s.io/api/autoscaling/v2beta1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -101,10 +100,7 @@ type kubeAPIServer struct {
 
 func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 	var (
-		needsVPA  = !k.values.Autoscaling.HVPAEnabled
 		needsHVPA = k.values.Autoscaling.HVPAEnabled && k.values.Autoscaling.Replicas != nil && *k.values.Autoscaling.Replicas > 0
-
-		vpaUpdateMode = autoscalingv1beta2.UpdateModeOff
 
 		horizontalPodAutoscaler = k.emptyHorizontalPodAutoscaler()
 		verticalPodAutoscaler   = k.emptyVerticalPodAutoscaler()
@@ -118,28 +114,8 @@ func (k *kubeAPIServer) Deploy(ctx context.Context) error {
 	if err := k.reconcileHorizontalPodAutoscaler(ctx, horizontalPodAutoscaler); err != nil {
 		return err
 	}
-
-	if needsVPA {
-		if _, err := controllerutils.GetAndCreateOrMergePatch(ctx, k.client, verticalPodAutoscaler, func() error {
-			verticalPodAutoscaler.Spec = autoscalingv1beta2.VerticalPodAutoscalerSpec{
-				TargetRef: &autoscalingv1.CrossVersionObjectReference{
-					APIVersion: appsv1.SchemeGroupVersion.String(),
-					Kind:       "Deployment",
-					// TODO: Replace with `deployment.Name` once the Deployment is generated by this component.
-					Name: v1beta1constants.DeploymentNameKubeAPIServer,
-				},
-				UpdatePolicy: &autoscalingv1beta2.PodUpdatePolicy{
-					UpdateMode: &vpaUpdateMode,
-				},
-			}
-			return nil
-		}); err != nil {
-			return err
-		}
-	} else {
-		if err := kutil.DeleteObject(ctx, k.client, verticalPodAutoscaler); err != nil {
-			return err
-		}
+	if err := k.reconcileVerticalPodAutoscaler(ctx, verticalPodAutoscaler); err != nil {
+		return err
 	}
 
 	if needsHVPA {
@@ -332,10 +308,6 @@ func (k *kubeAPIServer) GetValues() Values {
 }
 func (k *kubeAPIServer) SetAutoscalingReplicas(replicas *int32) {
 	k.values.Autoscaling.Replicas = replicas
-}
-
-func (k *kubeAPIServer) emptyVerticalPodAutoscaler() *autoscalingv1beta2.VerticalPodAutoscaler {
-	return &autoscalingv1beta2.VerticalPodAutoscaler{ObjectMeta: metav1.ObjectMeta{Name: v1beta1constants.DeploymentNameKubeAPIServer + "-vpa", Namespace: k.namespace}}
 }
 
 func (k *kubeAPIServer) emptyHVPA() *hvpav1alpha1.Hvpa {
